requirements.txt
trafilatura
aleph_alpha_client
rich
brave-search
types-requests
requests